Change url to api/register

+++ src/Core/UserBundle/Controller/UsersController.php
@@ -12,6 +12,11 @@ use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Sensio\Bundle\FrameworkExtraBundle\Configuration\ParamConverter;
 use FOS\RestBundle\Controller\Annotations\Post;
+use FOS\UserBundle\Event\FilterUserResponseEvent;
+use Symfony\Component\HttpFoundation\RedirectResponse;
+
+use Lexik\Bundle\JWTAuthenticationBundle\Event;
+use Symfony\Component\Security\Core\User\UserInterface;
 
 Class UsersController extends FOSRestController
 {
@@ -39,9 +44,9 @@ Class UsersController extends FOSRestController
     /**
      * @return array
      * @View()
-     * @Post("/registration")
+     * @Post("/register")
      */
-	public function registrationAction(Request $request)
+	public function registerAction(Request $request)
     {
 	    $formFactory = $this->container->get('fos_user.registration.form.factory');
 	    $userManager = $this->container->get('fos_user.user_manager');
@@ -63,29 +68,35 @@ Class UsersController extends FOSRestController
 	    if ('POST' === $request->getMethod()) {
             $form->bind($jsonPost);
 	        if ($form->isValid()) {
-                $username = $jsonPost['username'];
-                $password = $jsonPost['plainPassword']['first'];
-                $newUser = array(
-                    'username' => $username,
-                    'password' => $password
-                    );
-                $jsonLogin = json_encode($newUser);
+                // $username = $jsonPost['username'];
+                // $password = $jsonPost['plainPassword']['first'];
+                // $newUser = array(
+                //     'username' => $username,
+                //     'password' => $password
+                //     );
+                // $jsonLogin = json_encode($newUser);
 	            $event = new FormEvent($form, $request);
 	            $dispatcher->dispatch(FOSUserEvents::REGISTRATION_SUCCESS, $event);
 	            $userManager->updateUser($user);
-                $url = "http://localhost:8888/2SN_Core/web/app_dev.php/api/login_check";
-                $ch = curl_init($url);
-                curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "POST");
-                curl_setopt($ch, CURLOPT_POSTFIELDS, $jsonLogin);
-                curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
-                curl_setopt($ch, CURLOPT_HTTPHEADER, array(
-                    'Content-Type: application/json',
-                    'Content-Length: ' . strlen($jsonLogin))
-                );
-                $result = curl_exec($ch);
-	            $response = new Response($result);
+             //    $url = "http://localhost:8888/2SN_Core/web/app_dev.php/api/login_check";
+             //    $ch = curl_init($url);
+             //    curl_setopt($ch, CURLOPT_CUSTOMREQUEST, "POST");
+             //    curl_setopt($ch, CURLOPT_POSTFIELDS, $jsonLogin);
+             //    curl_setopt($ch, CURLOPT_RETURNTRANSFER, true);
+             //    curl_setopt($ch, CURLOPT_HTTPHEADER, array(
+             //        'Content-Type: application/json',
+             //        'Content-Length: ' . strlen($jsonLogin))
+             //    );
+             //    $result = curl_exec($ch);
+	            // $response = new Response($result);
+                if (null === $response = $event->getResponse()) {
+                    $url = $this->container->get('router')->generate('fos_user_registration_confirmed');
+                    $response = new RedirectResponse($url);
+                }
 
-	            return $response;
+                $dispatcher->dispatch(FOSUserEvents::REGISTRATION_COMPLETED, new FilterUserResponseEvent($user, $request, $response));
+
+                return $response;
 	        }
 	    }
         return array(400, $form);
@@ -98,6 +109,10 @@ Class UsersController extends FOSRestController
     public function deleteUserAction(User $user)
     {
         $em = $this->getDoctrine()->getManager();
+        $user = $entityManager->getRepository('CoreUserBundle:User')->find($id);
+        if (!$user) {
+            throw $this->createNotFoundException('User Not Found');
+        }
         $em->remove($user);
         $em->flush();
         $response = new Response("delete done", 200);
